add mbinfo

+++ include/multiboot.h
@@ -93,7 +93,7 @@
 
 typedef unsigned char           multiboot_uint8_t;
 typedef unsigned short          multiboot_uint16_t;
-typedef unsigned int            multiboot_uint32_t;
+typedef unsigned long int            multiboot_uint32_t;
 typedef unsigned long long      multiboot_uint64_t;
 
 struct multiboot_header
